requirements.txt
adjustText==0.8
boto3==1.22.13
cmake==3.22.4
contractions==0.1.73
cython==3.0.9
dask[complete]==2024.4.1
dataframe_image==0.2.3
dnspython==1.16.0
emoji==2.11.0
facebook-business==19.0.3
ftfy==6.2.0
google-ads==23.1.0
google-analytics-data==0.18.7
google-api-python-client==2.126.0
gunicorn==20.1.0
hdbscan==0.8.33
husl==4.0.3
ipython==7.29.0
jinja2==3.1.2
joblib==1.2.0
matplotlib==3.8.0
MiniSom==2.3.2
mlxtend==0.21.0
nbconvert==6.4.4
numpy==1.25.2
oauth2client==4.1.3
openpyxl==3.1.2
pandas==2.2.2
pdfminer.six==20231228
psycopg2-binary==2.9.9
pymongo==3.12.0
pymysql==1.0.2
python-dotenv==0.21.0
requests==2.31.0
s3fs==0.4.2
sacremoses==0.1.1
scikit-learn==1.2.0
scipy==1.9.3
seaborn==0.13.0
shap==0.41.0
simple-salesforce==1.12.5
sqlalchemy-bigquery==1.9.0
sqlalchemy-redshift==0.8.9
sqlalchemy==1.4.27
sshtunnel==0.4.0
statsmodels==0.13.5
stripe==5.0.0
tabula-py==2.9.0
tqdm==4.64.0
typing_extensions==4.5.0
umap==0.1.1
unidecode==1.3.4
uvicorn==0.18.3
xgboost==1.7.3
xlrd==2.0.1
